Fix vampire drain race conditions with atomic fixed-point energy
- Changed Agent.energy from f32 to u32 fixed-point (ENERGY_SCALE=100000) in both Rust and WGSL
- Updated agents_out/new_agents to use AgentAtomic struct with atomic<u32> energy field
- Inlined atomic operations at call sites (Naga doesn't allow ptr<storage,...> in function params)
- Vampire drain now uses atomic subtract/add to prevent energy creation via contention
- Victim loses 2x drained (atomically), attacker gains actual loss * 0.5
- Removed spatial grid lock/claim mechanism (redundant with atomic energy)
- Fixed snapshot/UI energy conversions to maintain f32 JSON compatibility
- Eliminated all illegal AgentAtomic struct copies (replaced with field-by-field writes)

+++ shaders/shared_types_only.wgsl
@@ -33,7 +33,7 @@ struct Agent {
     position: vec2<f32>,
     velocity: vec2<f32>,
     rotation: f32,
-    energy: f32,
+    energy: u32,
     energy_capacity: f32,
     torque_debug: f32,
     morphology_origin: vec2<f32>,
